+++ app/src/main/java/edu/gatech/cs2340/hkskh/Controllers/LoginActivity.java
@@ -1,4 +1,4 @@
-package edu.gatech.cs2340.hkskh.Controller;
+package edu.gatech.cs2340.hkskh.Controllers;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -19,7 +19,7 @@ public class LoginActivity extends AppCompatActivity {
 
         Button b1, b2;
         final EditText ed1, ed2;
-        ed1 = (EditText) findViewById(R.id.userText);
+        ed1 = findViewById(R.id.userText);
         ed2 = findViewById(R.id.passText);
         b1 = findViewById(R.id.logButton);
         b2 = findViewById(R.id.button);
