+++ app/src/main/java/edu/gatech/cs2340/hkskh/Models/User.java
@@ -1,4 +1,4 @@
-package edu.gatech.cs2340.hkskh.Model;
+package edu.gatech.cs2340.hkskh.Models;
 
 /**
  * Created by Kirby on 2/16/2018.
@@ -29,7 +29,6 @@ public class User {
         this("user", UserType.USER, "pass");
     }
 
-    //Getters ------------------------------------------------
     public String getName() {
         return name;
     }
@@ -41,6 +40,5 @@ public class User {
     public UserType getAccountType() {
         return type;
     }
-    //---------------------------------------------------------
 }
